backend/azure-functions/requirements.txt
azure-functions>=1.18.0
debugpy>=1.6.0